+++ src/OpenLoco/src/Jukebox.cpp
@@ -10,9 +10,8 @@ using namespace OpenLoco::Environment;
 
 namespace OpenLoco::Jukebox
 {
-    static MusicId currentTrack;  // 0x0050D434
-    static MusicId previousTrack; // 0x0050D435
-    static MusicId requestedNextTrack;
+    static MusicId selectedTrack; // 0x0050D434
+    static bool selectedTrackNotPlayedYet;
 
     // 0x004FE910
     static constexpr MusicInfo kMusicInfo[] = {
@@ -46,33 +45,32 @@ namespace OpenLoco::Jukebox
         { PathId::music_get_me_to_gladstone_bay, StringIds::music_get_me_to_gladstone_bay, 1918, 1926 },
         { PathId::music_sandy_track_blues, StringIds::music_sandy_track_blues, 1921, 1929 }
     };
+    static_assert(std::size(kMusicInfo) == kNumMusicTracks);
 
     const MusicInfo& getMusicInfo(MusicId track)
     {
         return kMusicInfo[track];
     }
 
+    // Note: this counts paused as playing
+    // TODO: there is no reason to check for kNoSong. Then this function can be deleted as it would be needless.
     bool isMusicPlaying()
     {
-        return (currentTrack != kNoSong && Config::get().audio.playJukeboxMusic);
+        return (selectedTrack != kNoSong && Config::get().audio.playJukeboxMusic);
     }
 
     MusicId getCurrentTrack()
     {
-        return currentTrack;
+        return selectedTrack;
     }
 
     StringId getSelectedTrackTitleId()
     {
-        if (requestedNextTrack != kNoSong)
+        if (selectedTrack == kNoSong)
         {
-            return kMusicInfo[requestedNextTrack].titleId;
+            return StringIds::music_none;
         }
-        if (currentTrack != kNoSong)
-        {
-            return kMusicInfo[currentTrack].titleId;
-        }
-        return StringIds::music_none;
+        return kMusicInfo[selectedTrack].titleId;
     }
 
     static void sortPlaylistByTitle(std::vector<uint8_t>& playlist, bool reversed = false)
@@ -191,10 +189,12 @@ namespace OpenLoco::Jukebox
         throw Exception::RuntimeError("Invalid MusicPlaylistType");
     }
 
-    static MusicId chooseNextMusicTrack(MusicId lastSong)
+    // Change selectedTrack to a random track from the selected playlist
+    static void chooseNextTrack()
     {
         auto playlist = makeSelectedPlaylist();
 
+        // Fallback playlists (TODO: does this even make sense? Surely the only danger to look out for is someone config editing their custom playlist)
         const auto& cfg = Config::get();
         if (playlist.empty() && cfg.audio.playlist != Config::MusicPlaylistType::currentEra)
         {
@@ -206,60 +206,58 @@ namespace OpenLoco::Jukebox
             playlist = makeAllMusicPlaylist();
         }
 
-        // Remove lastSong if it is present and not the only song, so that you do not get the same song twice.
-        // Assumes there is no more than one occurence of that song in the playlist.
+        // Unless it is the only track, prevent the same song from playing twice in a row.
+        // Assumes there is no more than one occurence of this track in the playlist.
         if (playlist.size() > 1)
         {
-            auto position = std::find(playlist.begin(), playlist.end(), lastSong);
+            auto position = std::find(playlist.begin(), playlist.end(), selectedTrack);
             if (position != playlist.end())
             {
                 playlist.erase(position);
             }
         }
 
+        // And pick a song!
         auto r = std::rand() % playlist.size();
-        auto track = playlist[r];
-        return track;
+        selectedTrack = playlist[r];
+
+        selectedTrackNotPlayedYet = true;
     }
 
-    const MusicInfo& changeTrack()
+    // If the next track has already been selected, mark it as played. Otherwise, choose the next track (and mark it as played).
+    // Returns the information for this track so that Audio.cpp can get its PathId to play.
+    const MusicInfo& consumeTrack()
     {
-        previousTrack = currentTrack;
-        if (requestedNextTrack != kNoSong)
+        if (!selectedTrackNotPlayedYet)
         {
-            currentTrack = requestedNextTrack;
-            requestedNextTrack = kNoSong;
+            chooseNextTrack();
         }
-        else
-        {
-            currentTrack = chooseNextMusicTrack(previousTrack);
-        }
+        selectedTrackNotPlayedYet = false;
 
-        // Return the information for this track so that Audio.cpp can get its PathId to play.
-        return kMusicInfo[currentTrack];
+        return kMusicInfo[selectedTrack];
     }
 
     // The player manually selects a song from the drop-down in the music options.
     bool requestTrack(MusicId track)
     {
-        if (track == currentTrack)
+        assert(track < kNumMusicTracks); // Also catches kNoSong ("[None]"), which is impossible to request.
+
+        if (track == selectedTrack)
         {
             return false;
         }
 
-        requestedNextTrack = track;
+        selectedTrack = track;
+        selectedTrackNotPlayedYet = true;
 
-        // Stop the currently playing song prematurely
+        // If a song is playing, stop it prematurely.
         Audio::stopMusic();
 
         // Previously 0x0050D430 '_songProgress' would be set to 0 here, but that loco global was no longer used for anything.
 
-        assert(requestedNextTrack != kNoSong); // "[None]" should not appear in the drop-down, how did you request it?
-
         return true;
     }
 
-    // Prematurely stops the current song so that another can be played.
     bool skipCurrentTrack()
     {
         if (Config::get().audio.playJukeboxMusic == 0)
@@ -267,8 +265,10 @@ namespace OpenLoco::Jukebox
             return false;
         }
 
-        // By stopping the music, the next time OpenLoco::tick() calls Audio::playBackgroundMusic(),
-        // it will detect that the channel is not playing anything, and so it will play the next music track.
+        // Changing the track here makes the skip button responsive even when paused.
+        chooseNextTrack();
+
+        // Stopping the music causes Audio::playBackgroundMusic() to call Jukebox::consumeTrack() next time it is called.
         Audio::stopMusic();
 
         return true;
@@ -288,7 +288,7 @@ namespace OpenLoco::Jukebox
         Config::write();
 
         Audio::stopMusic();
-        currentTrack = kNoSong;
+        selectedTrack = kNoSong;
 
         return true;
     }
@@ -305,13 +305,14 @@ namespace OpenLoco::Jukebox
         cfg.playJukeboxMusic = 1;
         Config::write();
 
+        chooseNextTrack();
+
         return true;
     }
 
     void resetJukebox()
     {
-        currentTrack = kNoSong;
-        previousTrack = kNoSong;
-        requestedNextTrack = kNoSong;
+        selectedTrack = kNoSong;
+        selectedTrackNotPlayedYet = false;
     }
 }